Page: Change the `application` property to String.
The previous implementation was rather error-prone for users of the
library. E.g. the following page did not actually set the application:

```
class TestPage: Page {
    let application = "com.apple.springboard"

    var body: PageDescription { StaticText("Test") }
}
```

The expected type in the `Page` protocol is `String?`. But the inferred
type in the declaration above is `String`.

The new implementation will make this case more straight forward for
users of the framework and will avoid errors which are rather hard to
debug for users without deep knowledge of the insights of the framework.