Reorganization of the code.
Each page element now has its own file, which will enable us to
add element-specific methods and functionality in the future.

It also broke up the big `PageElement.swift` file that contained
too many things and that was rather confusing. Some of the other
big files were split as well.

Additionally, all structures which are not meant to be used by
end users because they are not part of the public API and simply
needed to make the public API work were moved into a separated
directory.